+++ aten/src/THCUNN/BatchNormalization.cu
@@ -7,15 +7,15 @@
 #include "THCDeviceTensor.cuh"
 #include "THCDeviceTensorUtils.cuh"
 #include "THCDeviceUtils.cuh"
-const int WARP_SIZE = 32;
+const int64_t WARP_SIZE = 32;
 
 // The maximum number of threads in a block
-const int MAX_BLOCK_SIZE = 512;
+const int64_t MAX_BLOCK_SIZE = 512;
 
 // Number of threads in a block given an input size up to MAX_BLOCK_SIZE
-static int getNumThreads(int nElem) {
-  int threadSizes[5] = { 32, 64, 128, 256, MAX_BLOCK_SIZE };
-  for (int i = 0; i != 5; ++i) {
+static int64_t getNumThreads(int64_t nElem) {
+  int64_t threadSizes[5] = { 32, 64, 128, 256, MAX_BLOCK_SIZE };
+  for (int64_t i = 0; i != 5; ++i) {
     if (nElem <= threadSizes[i]) {
       return threadSizes[i];
     }
@@ -24,7 +24,7 @@ static int getNumThreads(int nElem) {
 }
 
 // Returns the index of the most significant 1 bit in `val`.
-__device__ __forceinline__ int getMSB(int val) {
+__device__ __forceinline__ int64_t getMSB(int64_t val) {
   return 31 - __clz(val);
 }
 
@@ -34,7 +34,7 @@ struct Float2 {
   __device__ Float2() {}
   __device__ Float2(Dtype v1, Dtype v2) : v1(ScalarConvert<Dtype, Acctype>::to(v1)), v2(ScalarConvert<Dtype, Acctype>::to(v2)) {}
   __device__ Float2(Dtype v) : v1(ScalarConvert<Dtype, Acctype>::to(v)), v2(ScalarConvert<Dtype, Acctype>::to(v)) {}
-  __device__ Float2(int v) : v1(ScalarConvert<int, Acctype>::to(v)), v2(ScalarConvert<int, Acctype>::to(v)) {}
+  __device__ Float2(int64_t v) : v1(ScalarConvert<int64_t, Acctype>::to(v)), v2(ScalarConvert<int64_t, Acctype>::to(v)) {}
   __device__ Float2& operator+=(const Float2& a) {
     v1 += a.v1;
     v2 += a.v2;
@@ -45,7 +45,7 @@ struct Float2 {
 template <typename Dtype, typename Acctype, typename DeviceTensor3>
 struct SumOp {
   __device__ SumOp(const DeviceTensor3 t) : tensor(t) {}
-  __device__ __forceinline__ Acctype operator()(int batch, int plane, int n) {
+  __device__ __forceinline__ Acctype operator()(int64_t batch, int64_t plane, int64_t n) {
     return ScalarConvert<Dtype, Acctype>::to(tensor[batch][plane][n]);
   }
   const DeviceTensor3 tensor;
@@ -54,7 +54,7 @@ struct SumOp {
 template <typename Dtype, typename Acctype, typename DeviceTensor3>
 struct VarOp {
   __device__ VarOp(Acctype m, const DeviceTensor3 t) : mean(m), tensor(t) {}
-  __device__ __forceinline__ Acctype operator()(int batch, int plane, int n) {
+  __device__ __forceinline__ Acctype operator()(int64_t batch, int64_t plane, int64_t n) {
     Dtype val = tensor[batch][plane][n];
     return (val - mean) * (val - mean);
   }
@@ -66,7 +66,7 @@ template <typename Dtype, typename Acctype, typename DeviceTensor3>
 struct GradOp {
   __device__ GradOp(Acctype m, const DeviceTensor3 i, const DeviceTensor3 g)
     : mean(m), input(i), gradOutput(g) {}
-  __device__ __forceinline__ Float2<Dtype, Acctype> operator()(int batch, int plane, int n) {
+  __device__ __forceinline__ Float2<Dtype, Acctype> operator()(int64_t batch, int64_t plane, int64_t n) {
     Dtype g = gradOutput[batch][plane][n];
     Dtype c = ScalarConvert<Acctype, Dtype>::to(input[batch][plane][n] - mean);
     return Float2<Dtype, Acctype>(g, g * c);
@@ -80,15 +80,15 @@ struct GradOp {
 template <typename T>
 static __device__ __forceinline__ T warpSum(T val) {
 #if __CUDA_ARCH__ >= 300
-  for (int i = 0; i < getMSB(WARP_SIZE); ++i) {
+  for (int64_t i = 0; i < getMSB(WARP_SIZE); ++i) {
     val += WARP_SHFL_XOR(val, 1 << i, WARP_SIZE);
   }
 #else
   __shared__ T values[MAX_BLOCK_SIZE];
   values[threadIdx.x] = val;
   __threadfence_block();
-  const int base = (threadIdx.x / WARP_SIZE) * WARP_SIZE;
-  for (int i = 1; i < WARP_SIZE; i++) {
+  const int64_t base = (threadIdx.x / WARP_SIZE) * WARP_SIZE;
+  for (int64_t i = 1; i < WARP_SIZE; i++) {
     val += values[base + ((i + threadIdx.x) % WARP_SIZE)];
   }
 #endif
@@ -104,10 +104,10 @@ static __device__ __forceinline__ Float2<Dtype, Acctype> warpSum(Float2<Dtype, A
 
 // Sum across (batch, x/y/z) applying Op() pointwise
 template<typename T, typename Op, typename DeviceTensor3>
-__device__ T reduce(Op op, DeviceTensor3 tensor, int plane) {
+__device__ T reduce(Op op, DeviceTensor3 tensor, int64_t plane) {
   T sum = (T)0;
-  for (int batch = 0; batch < tensor.getSize(0); ++batch) {
-    for (int x = threadIdx.x; x < tensor.getSize(2); x += blockDim.x) {
+  for (int64_t batch = 0; batch < tensor.getSize(0); ++batch) {
+    for (int64_t x = threadIdx.x; x < tensor.getSize(2); x += blockDim.x) {
       sum += op(batch, plane, x);
     }
   }
@@ -148,7 +148,7 @@ __global__ void BatchNormalizationUpdateOutputInference_kernel(
     const DeviceTensor1 bias,
     Acctype epsilon) {
 
-  int plane = blockIdx.x;
+  int64_t plane = blockIdx.x;
 
   Acctype invstd = Acctype(1) / sqrt(runningVar[plane].ldg() + epsilon);
   Acctype mean = ScalarConvert<Dtype, Acctype>::to(runningMean[plane].ldg());
@@ -156,8 +156,8 @@ __global__ void BatchNormalizationUpdateOutputInference_kernel(
   Acctype beta = bias.numElements() > 0 ? ScalarConvert<Dtype, Acctype>::to(bias[plane].ldg()) : Acctype(0);
 
   // Write normalized and update the output
-  for (int batch = 0; batch < input.getSize(0); batch++) {
-    for (int x = threadIdx.x; x < input.getSize(2); x += blockDim.x) {
+  for (int64_t batch = 0; batch < input.getSize(0); batch++) {
+    for (int64_t x = threadIdx.x; x < input.getSize(2); x += blockDim.x) {
       Dtype inp = input[batch][plane][x].ldg();
       output[batch][plane][x] = ScalarConvert<Acctype, Dtype>::to(gamma * (inp - mean) * invstd + beta);
     }
@@ -177,8 +177,8 @@ __global__ void BatchNormalizationUpdateOutput_kernel(
     DeviceTensor1 saveMean,
     DeviceTensor1 saveStd) {
 
-  int plane = blockIdx.x;
-  int N = input.getSize(0) * input.getSize(2);
+  int64_t plane = blockIdx.x;
+  int64_t N = input.getSize(0) * input.getSize(2);
 
   Acctype norm = Acctype(1) / N;
 
@@ -208,8 +208,8 @@ __global__ void BatchNormalizationUpdateOutput_kernel(
   // Write normalized and update the output
   Acctype gamma = weight.numElements() > 0 ? ScalarConvert<Dtype, Acctype>::to(weight[plane]) : ScalarConvert<int, Acctype>::to(1);
   Acctype beta = bias.numElements() > 0 ? ScalarConvert<Dtype, Acctype>::to(bias[plane]) : ScalarConvert<int, Acctype>::to(0);
-  for (int batch = 0; batch < input.getSize(0); ++batch) {
-    for (int x = threadIdx.x; x < input.getSize(2); x += blockDim.x) {
+  for (int64_t batch = 0; batch < input.getSize(0); ++batch) {
+    for (int64_t x = threadIdx.x; x < input.getSize(2); x += blockDim.x) {
       Dtype inp = input[batch][plane][x].ldg();
       output[batch][plane][x] = ScalarConvert<Acctype, Dtype>::to(gamma * (inp - mean) * invStd + beta);
     }
@@ -232,8 +232,8 @@ __global__ void BatchNormalizationBackward_kernel(
     Acctype scale,
     double eps) {
 
-  int plane = blockIdx.x;
-  int N = gradOutput.getSize(0) * gradOutput.getSize(2);
+  int64_t plane = blockIdx.x;
+  int64_t N = gradOutput.getSize(0) * gradOutput.getSize(2);
 
   Acctype mean, stdVal;
   if (train) {
@@ -260,8 +260,8 @@ __global__ void BatchNormalizationBackward_kernel(
   Acctype gradScale = stdVal * weightVal;
 
   if (gradInput.numElements() > 0) {
-    for (int batch = 0; batch < gradOutput.getSize(0); ++batch) {
-      for (int x = threadIdx.x; x < gradOutput.getSize(2); x += blockDim.x) {
+    for (int64_t batch = 0; batch < gradOutput.getSize(0); ++batch) {
+      for (int64_t x = threadIdx.x; x < gradOutput.getSize(2); x += blockDim.x) {
         Dtype gradOut = gradOutput[batch][plane][x];
         if (train) {
           Dtype inp = input[batch][plane][x];
